file structure reorganization
