+++ js/screens/game-3.js
@@ -1,11 +1,11 @@
 import {
   makeElement,
   showScreen,
-} from "./utils";
+} from "../utils";
 import stats from './stats';
 import greeting from "./greeting";
-import statsTemplate from './stats-template';
-import headerTemplate from "./header/header-template";
+import statsTemplate from '../templates/stats-template';
+import headerTemplate from "../templates/header-template";
 
 const game3 = (state) => {
   const gameTask = `<p class="game__task">${state.GAME_TASKS.threePic}</p>`;
