+++ js/templates/stats-template.js
